refactor(kb): extract 7 state components from layout into separate files
The KB layout grew to ~330 lines with 7 inline component definitions.
Extract LoadingSkeleton, WorkspaceNotReady, NoProjectState, UnknownError,
EmptyState, DesktopPlaceholder, and KbErrorBoundary into individual files
under components/kb/ with a barrel index for clean imports.
