+++ apps/web-platform/components/kb/index.ts
@@ -1,0 +1,7 @@
+export { DesktopPlaceholder } from "./desktop-placeholder";
+export { EmptyState } from "./empty-state";
+export { KbErrorBoundary } from "./kb-error-boundary";
+export { LoadingSkeleton } from "./loading-skeleton";
+export { NoProjectState } from "./no-project-state";
+export { UnknownError } from "./unknown-error";
+export { WorkspaceNotReady } from "./workspace-not-ready";
